requirements.txt
annotated-types==0.7.0
asttokens==2.4.1
cachetools==5.4.0
certifi==2024.7.4
charset-normalizer==3.3.2
decorator==5.1.1
executing==2.0.1
google-ai-generativelanguage==0.6.6
google-api-core==2.19.1
google-api-python-client==2.138.0
google-auth==2.32.0
google-auth-httplib2==0.2.0
google-generativeai==0.7.2
googleapis-common-protos==1.63.2
httplib2==0.22.0
idna==3.7
ipython==8.26.0
jedi==0.19.1
markdown-it-py==3.0.0
matplotlib-inline==0.1.7
mdurl==0.1.2
parso==0.8.4
pexpect==4.9.0
prompt_toolkit==3.0.47
proto-plus==1.24.0
protobuf==4.25.4
ptyprocess==0.7.0
pure_eval==0.2.3
pyasn1==0.6.0
pyasn1_modules==0.4.0
pydantic==2.8.2
pydantic_core==2.20.1
Pygments==2.18.0
pyparsing==3.1.2
requests==2.32.3
rich==13.7.1
rsa==4.9
six==1.16.0
stack-data==0.6.3
tqdm==4.66.4
traitlets==5.14.3
typing_extensions==4.12.2
uritemplate==4.1.1
urllib3==2.2.2
wcwidth==0.2.13
